requirements.txt
openshift
pyopenssl
